# api-sigea/openapi/sigec.yaml
openapi: 3.0.0
info:
  title: SigecAPIDMS
  description: Rest Api che consente la gestione e il monitoraggio degli eventi della regione Puglia
  version: v1
servers:
  - url: "http://localhost:9292/SigecAPI"
    description: LOCALE
  - url: ND
    description: SVILUPPO
  - url: ND
    description: COLLAUDO
  - url: ND
    description: PRODUZIONE
paths:
...
  /segnalazioni:
    post:
      tags:
        - Endpoint accessibili solo da SIGEA
      summary: "API per l'inserimento di una segnalazione evento"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SegnalazioneModel"
      operationId: insertSegnalazione
      responses:
        "200":
          content: {}
          description: Segnalazione correttamente creata
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
    get:
      tags:
        - Endpoint accessibili solo da SIGEA
      summary: API per la lettura della lista delle segnalazioni evento in un dato comune e in un dato giorno
      parameters:
        - name: dataDa
          in: query
          description: Data in cui effettuare la ricerca
          required: true
          schema:
            type: string
            format: date
        - name: codIstat
          in: query
          description: Codice istat del comune su cui effettuare la ricerca
          required: true
          schema:
            type: string
      operationId: getListaSegnalazioni
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SegnalazioneModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
  "/segnalazioni/{id}":
    get:
      tags:
        - SegnalazioneAPI
      summary: API per la lettura di una segnalazione evento
      parameters:
        - name: id
          in: path
          description: Id della segnalazione su cui effettuare la lettura
          required: true
          schema:
            type: number
            format: int64
      operationId: getSegnalazione
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegnalazioneModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
    put:
      tags:
        - Endpoint accessibili solo da SIGEA
      summary: API per la modifica dello stato di una segnalazione evento
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SegnalazioneStatoPatchModel"
      parameters:
        - name: id
          in: path
          description: Id della segnalazione su cui effettuare il cambio di stato
          required: true
          schema:
            type: integer
            format: int64
      operationId: cambiaStato
      responses:
        "200":
          description: Cambio stato avvenuto correttamente
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
  /segnalazioni/listapaginata:
    get:
      tags:
        - SegnalazioneAPI
      summary: API per la lettura della lista paginata delle segnalazioni evento (gli oggetti presenti nella lista ottenuta sono una proiezione del model rappresentante la segnalazione vera e propria)
      parameters:
        - name: wrappedFilter
          in: query
          description: filtro codificato da applicare per la paginazione
          required: true
          schema:
            type: object
            properties:
              detail:
                $ref: "#/components/schemas/WrapperFilterPaginationDetail"
              filter:
                $ref: "#/components/schemas/SegnalazioneFilter"
      operationId: getListaPaginataSegnalazioni
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  draw:
                    type: integer
                  recordsTotal:
                    type: integer
                  recordsFiltered:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SegnalazioneModelList"
          description: Oggetto contenente le specifiche di paginazione
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
...
  /utenti:
    put:
      tags:
        - Endpoint accessibili solo da SIGEA
      summary: "API per la profilazione dell'utente, i dati vengono estratti direttamente dal token"
      operationId: saveUtente
      parameters:
        - name: tipologia
          in: query
          description: "id tipologia da associare all'utente"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profilazione avvenuta correttamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DettaglioUtenteModel"
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
  /tipologie:
    get:
      tags:
        - UtenteAPI
      summary: API per la lettura delle possibili tipologie utente
      operationId: getTipologie
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TipologiaModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
  /redazioni:
    get:
      tags:
        - UtenteAPI
      summary: API per la lettura delle redazioni
      operationId: getRedazioni
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RedazioneModel"
          description: Oggetto contenente le specifiche
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
...
  /riconciliazioni:
    get:
      tags:
        - Endpoint accessibili solo dal DMS
      summary: "API per la lettura dell'elenco degli eventi da riconciliare per una o più email"
      operationId: riconciliazioneElenco
      parameters:
        - name: elencomail
          in: query
          description: elenco delle emaill
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ElencoEventiEmail"
          description: "Array contenente l'elenco degli eventi da riconciliare per ogni email"
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
  /riconciliazioni/associazioni:
    post:
      tags:
        - Endpoint accessibili solo dal DMS
      summary: API per effetuare la riconciliazione sugli eventi
      operationId: riconciliazioneAssociazione
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ElencoRiconciliazioni"
      responses:
        "200":
          description: Riconciliazione completata
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "409":
          description: Riconciliazione non effettuata a causa di un conflitto su uno o più eventi
          content:
            application/json:
              schema:
                type: object
                description: "id dell'evento su cui si è verificato il conflitto"
                additionalProperties:
                  type: string
                  description: "Titolo italiano dell'evento"
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "503":
          $ref: "#/components/responses/503ServiceUnavailable"
        default:
          $ref: "#/components/responses/default"
      security:
        - bearerAuth: []
      x-auth-type: "Application & Application User"
      x-throttling-tier: Unlimited
...
components:
  schemas:
...
    DettaglioUtenteModel:
      type: object
      properties:
        utenteId:
          type: number
          format: int64
          description: Id utente generato dal DMS
        username:
          type: string
          description: Username
        nome:
          type: string
          description: Nome
        cognome:
          type: string
          description: Cognome
        codFiscale:
          type: string
          description: Codice fiscale
        email:
          type: string
          description: E-mail
        tel:
          type: string
          description: Telefono
        cel:
          type: string
          description: Cellulare
        tipologia:
          $ref: "#/components/schemas/TipologiaModel"
...
    ElencoEventiEmail:
      type: object
      properties:
        email:
          type: string
          description: Email a cui è agganciata la lista di eventi
        eventi:
          type: object
          description: "id dell'evento"
          additionalProperties:
            type: string
            description: "Titolo italiano dell'evento"
        motivazioneErrore:
          type: string
          description: "eventuale motivazione dell'errore"
    ElencoRiconciliazioni:
      type: object
      properties:
        idAttivita:
          type: number
          format: int64
          description: "id dell'attività su cui effettuare la riconciliazione"
        denominazioneAttivita:
          type: string
          description: "denominazione dell'attività su cui effettuare la riconciliazione"
        eventi:
          type: object
          description: "id dell'evento"
          additionalProperties:
            type: string
            description: "Titolo italiano dell'evento"
...
    RedazioneModel:
      type: object
      properties:
        redazioneId:
          type: string
          description: Id redazione
        nome:
          type: string
          description: Nome redazione
        link:
          type: string
          description: Link alla home del front office su cui sono visibili le pubblicazioni
        jspName:
          type: string
          description: Nome della jsp contenente la scheda di redazione qualora sia stata implementata su SigeaWeb
        linkScheda:
          type: string
          description: Link alla scheda di redazione qualora sia stata implementata su di un modulo esterno a SigeaWeb
        autoSpubblicazione:
          type: boolean
          description: "Indica se la redazione prevede l'autospubblicazione"
        attivazioneJMS:
          type: boolean
          description: "Indica se la redazione prevede l'uso del JMS"
...
    RuoloModel:
      type: object
      properties:
        ruoloId:
          type: string
          description: Id del ruolo
        permessi:
          type: object
          description: Id del permesso
          additionalProperties:
            type: boolean
            description: Indica se il ruolo ha il premesso attivo
    SegnalazioneFilter:
      type: object
      properties:
        idUtenteIns:
          type: number
          format: int64
          description: Id utente che ha creato la segnalazione
        titolo:
          type: string
          description: Titolo
        dataDa:
          type: string
          description: Data da
        dataA:
          type: string
          description: Data a
        status:
          type: string
          description: Stato
        codIstat:
          type: string
          description: Codice del comune proveniente dal modulo dei dati territoriali
        comuneEstero:
          type: string
          description: Comune estero
    SegnalazioneModel:
      type: object
      properties:
        segnalazioneId:
          type: number
          format: int64
          description: Id della segnalazione
        nome:
          type: string
          description: Titolo della segnalazione
        descrizione:
          type: string
          description: Descrizione della segnalazione
        dataDa:
          type: string
          format: date
          description: Data da
        dataA:
          type: string
          format: date
          description: Data a
        location:
          type: string
          description: Nome della location
        indirizzo:
          type: string
          description: Indirizzo
        riferimento:
          type: string
          description: Riferimento
        utenteId:
          type: number
          format: int64
          description: Id utente che crea la segnalazione
        codIstat:
          type: string
          description: Codice del comune proveniente dal modulo dei dati territoriali
        comune:
          type: string
          description: Nome del comune proveniente dal modulo dei dati territoriali
        comuneEstero:
          type: string
          description: Comune estero
    SegnalazioneModelList:
      type: object
      properties:
        segnalazioneId:
          type: string
          description: Id della segnalazione
        nome:
          type: string
          description: Nome della segnalazione
        dataDa:
          type: string
          description: Data da
        dataA:
          type: string
          description: Data a
        comune:
          type: string
          description: Comune
        riferimento:
          type: string
          description: Riferimento
        dataIns:
          type: string
          description: Data di creazione
        status:
          type: string
          description: Stato corrente
    SegnalazioneStatoPatchModel:
      type: object
      properties:
        stato:
          type: string
          description: Stato aggiornato
...
    TipologiaModel:
      type: object
      properties:
        tipologiaId:
          type: string
          description: Id della tipologia
        descrizione:
          type: string
          description: Descrizione della tipologia
        isRedattore:
          type: boolean
          description: Flag che indica se si tratta di un profilo redazionale
        ruoliSet:
          type: array
          description: Lista dei ruoli assunti da questa tipologia
          items:
            $ref: "#/components/schemas/RuoloModel"
        redazioniSet:
          type: array
          description: Lista di redazioni collegata a questa tipologia
          items:
            $ref: "#/components/schemas/RedazioneModel"
...
    WrapperFilterPaginationDetail:
      type: object
      properties:
        draw:
          type: number
          format: int32
          description: "Numero incrementale che definisce l'attuale richiesta (utilizzato a scopo di sicurezza)"
        start:
          type: number
          format: int32
          description: Risultato di partenza della paginazione
        length:
          type: number
          format: int32
          description: Lunghezza di paginazione
        search:
          type: string
          description: Stringa di ricerca testuale
        order:
          type: string
          description: "Colonna per la quale effettuare l'ordinamento"
        dir:
          type: string
          description: "Direzione dell'ordinamento"
    Problem:
      properties:
        detail:
          description: "A human readable explanation specific to this occurrence of the\nproblem. You MUST NOT expose internal informations, personal\ndata or implementation details through this field.\n"
          type: string
        instance:
          description: "An absolute URI that identifies the specific occurrence of the problem.\nIt may or may not yield further information if dereferenced.\n"
          format: uri
          type: string
        status:
          description: "The HTTP status code generated by the origin server for this occurrence\nof the problem.\n"
          exclusiveMaximum: true
          format: int32
          maximum: 600
          minimum: 100
          type: integer
        title:
          description: "A short, summary of the problem type. Written in english and readable\nfor engineers (usually not suited for non technical stakeholders and\nnot localized); example: Service Unavailable\n"
          type: string
        type:
          default: "about:blank"
          description: "An absolute URI that identifies the problem type.  When dereferenced,\nit SHOULD provide human-readable documentation for the problem type\n(e.g., using HTML).\n"
          format: uri
          type: string
      type: object
  responses:
...
    400BadRequest:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Bad Request
...
    404NotFound:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Not Found
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
...
    503ServiceUnavailable:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Service Unavailable
      headers:
        Retry-After:
          $ref: "#/components/headers/Retry-After"
    401Unauthorized:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Unauthorized
    default:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Unexpected error
  headers:
    Retry-After:
      description: "Retry contacting the endpoint *at least* after seconds.\nSee https://tools.ietf.org/html/rfc7231#section-7.1.3"
      schema:
        format: int32
        type: integer
    X-RateLimit-Limit:
      description: The number of allowed requests in the current period
      schema:
        format: int32
        type: integer
    X-RateLimit-Remaining:
      description: The number of remaining requests in the current period
      schema:
        format: int32
        type: integer
    X-RateLimit-Reset:
      description: The number of seconds left in the current period
      schema:
        format: int32
        type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer